refactor(encryption): rename src/encryption to src/crypto
